modularize the udma module
